Find Common Muliple
Write a function hasSmallCommonMultiple that returns true if its two integer input arguments arg1, arg2 have a common multiple that is less than arg1 * arg2. For example, on input 10, 15, the function returns true, since 30 is a common multiple of 10,15 that is less than 10*15 = 150. On the other hand, the function returns false on input 8, 9 since the smallest common multiple for these is 72.